Removed a number of dependencies which won't be used
Modified the fig 5 gene list function to output all genes DE in at least one RNA-seq condition as well as the specific list used for the heatmap.

Completed a function to extract GO annotations just for RNA-seq DE genes (to reduce the size of the temporary file stored on GitHub)

Added a function to extract a list of TFs from AnimalTFDB reference files

Fixed the formatting of Fig 5

Added in the GO membership & annotation tables back into Fig. 5b